+++ plugins/kuke-ai-agent/plugin.json
@@ -1,7 +1,7 @@
 {
   "name": "kuke-ai-agent",
   "title": "Kuke AI Agent",
-  "description": "本地全能 AI Agent，支持 读写本地文件、执行终端命令、联网搜索、Todo；具备强大的 AI Coding 能力、美观的UI和动效、多种完善的安全机制、审查模式，可自由模型配置，支持主流 OpenAI 兼容模型自由切换。",
+  "description": "本地全能 AI Agent，支持读写本地文件、执行终端命令、联网搜索、Todo；具备强大的 AI Coding 能力，支持 OpenAI / Anthropic (Claude) 多种模型自由切换，美观的 UI 和动效，多种安全机制。",
   "version": "1.0.0",
   "author": "kukemc",
   "homepage": "https://github.com/kukemc",
